requirements.txt
flask
pandas
numpy
scikit-learn
plotly
openpyxl
dvc
python-dotenv
gunicorn
pytest 
gdrive
dvc[gdrive]